requirements.txt
pandas
streamlit
matplotlib
seaborn
mysql-connector-python
